requirements.txt
pydantic==1.9.0
torch==1.11.0
torchvision==0.12.0
Pillow==9.0.1

numpy<2.0
fastapi==0.78.0
